Develop

+++ controllers/api/comment.js
@@ -13,12 +13,12 @@ router.get("/", (req, res) => {
 router.post("/", (req, res) => {
   // expects => {comment_text: "This is the comment", user_id: 1, post_id: 2}
 
-  // removed the session info so that routes will work while testing TODO: add back in session check
-  // if (req.session) {
+  
+  if (req.session) {
     Comment.create({
       comment_text: req.body.comment_text,
 
-      // user_id: req.session.user_id,
+      user_id: req.session.user_id,
       user_id: req.body.user_id,
       post_id: req.body.post_id,
     })
@@ -27,13 +27,13 @@ router.post("/", (req, res) => {
         console.log(err);
         res.status(400).json(err);
       });
-  // }
+  }
 });
 
 router.delete("/:id", (req, res) => {
 
-  //removed check for session so that the routes work. TODO: uncomment the session code 
-  // if (req.session) {
+  
+  if (req.session) {
     Comment.destroy({
       where: {
         id: req.params.id,
@@ -50,7 +50,7 @@ router.delete("/:id", (req, res) => {
         console.log(err);
         res.status(500).json(err);
       });
-  // }
+  }
 });
 
 module.exports = router;
